Split out test_engine_close_spider.py.
